requirements.txt
torch>=1.9.0
scikit-learn>=0.24.0
Pillow>=8.0.0
PyYAML>=5.4.0
tensorboard>=2.5.0
